# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DataFlux API
  description: |
    DataFlux ist eine Universal AI-native Database für Medieninhalte mit einer Plugin-Architektur und Multi-Modal Search Capabilities.
    
    ## Features
    - **Multi-Modal Search**: Suche über Video, Audio, Bilder und Dokumente
    - **AI-Powered Analysis**: Automatische Inhaltsanalyse mit Machine Learning
    - **Real-time Processing**: Streaming und Batch-Verarbeitung
    - **Scalable Architecture**: Microservices-basierte Architektur
    - **Enterprise Security**: JWT Authentication und RBAC
    
    ## Authentication
    Alle API-Endpunkte (außer `/health` und `/docs`) erfordern Authentifizierung über JWT-Token.
    
    ```bash
    # Login
    curl -X POST http://localhost:8006/auth/login \
      -H "Content-Type: application/json" \
      -d '{"username":"admin","password":"admin123"}'
    
    # Use token in requests
    curl -H "Authorization: Bearer <token>" \
      http://localhost:8003/api/v1/search
    ```
  version: 1.0.0
  contact:
    name: DataFlux Support
    email: support@dataflux.com
    url: https://dataflux.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.dataflux.com
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and receive JWT tokens
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
              required: [username, password]
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  token_type:
                    type: string
                    example: bearer
                  refresh_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      tags: [Authentication]
      summary: User registration
      description: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      description: Get current user information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Asset endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserCreate:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: john_doe
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          minLength: 8
          example: securepassword123
        full_name:
          type: string
          example: John Doe
        role:
          type: string
          enum: [user, analyst, admin]
          default: user
      required: [username, email, password]
    UserResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...